refactor: 헤더 로딩 수정
- Providers 파일 폴더 위치 변경
