+++ src/app/layout.tsx
@@ -7,15 +7,19 @@ import Script from 'next/script';
 import { usePathname } from 'next/navigation';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import LocalStorage from '@/utils/localstorage';
 import Header from './components/Header';
 import Footer from './components/Footer';
-import Providers from '../utils/providers';
+import Providers from './components/Providers';
 
 const inter = Inter({ subsets: ['latin'] });
 
 const queryClient = new QueryClient();
 
 const RootLayout = ({ children }: { children: React.ReactNode }) => {
+  if (LocalStorage.getItem('isLoggedIn') === null) {
+    LocalStorage.setItem('isLoggedIn', 'false');
+  }
   const pathname = usePathname();
 
   const withoutHeader =
